refactor: update file structure and delete commented code
