+++ src/App.tsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { BrowserRouter as Router } from "react-router-dom";
 import "./App.css";
 import AppRoutes from "./components/app-routes/app-routes";
-import ErrorBoundary from "./components/error-boundary-component/error-boundary-component";
+import ErrorBoundary from "./components/error-boundary/error-boundary";
 import { ThemeContext } from "./contexts/themeContext";
 
 const App: React.FC = () => {
