+++ src/components/app-routes/app-routes.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
 import ErrorPage from "../../views/error-page/error-page";
-import Main from "../../views/main/main-page";
+import Main from "../../views/main-page/main-page";
 
 const AppRoutes: React.FC = () => {
   return (
