+++ src/components/details/details.module.css
