+++ src/components/details/planet/planet.module.css
