+++ src/components/details/starship/starship.module.css
