+++ src/components/details/vehicles/vehicles.module.css
