+++ src/views/main-page/main-page.test.tsx
